README.md
A user has committed cyber crimes and lucky for me I know just the person to send digging through EVERYTHING on this dark web site that fronts as having standards and following the law. He has limitless contacts and tools at his disposal due to his specialty being cyber crimes and his high government clearance. I already had an ongoing local investigation due to stalking and deadly threats against me and now it has escalated. This sites coding has put my children even more at risk and now the FBI, my contact and local authorities will be given all evidence. It is in very safe hands. A deal can be made if I am contacted very quickly otherwise don't think being based in India will stop them from digging through every inch and thread of this site and all users involved. For sure the user Daemon is involved to some extent. Or at least his coding. They are active on this device or someone is but Daemon programming specifically python and Ruby were embedded in the phone that is gonna be handed over today. The list of crimes committed is endless......all FEDERAL level. Impersonation, theft of intellectual propert, fraud, identity theft, child porn and much more. I have been warned by officials of blackmail or public releasing of these things but I have much more they are interested in from this site as the user left it open in my last phone that was hacked. The VA is involved as well due to his mental state, if it is the same person as the stalker. The clock is ticking